requirements.txt
numpy
pandas
matplotlib
seaborn
scipy
pytest
flask
requests
beautifulsoup4
nltk
tensorflow-cpu
torch
torchvision
black
pylint
ipython
flask-cors
python-dotenv
pyflakes
pydocstyle
pygments
watchdog
psutil 